refactoring to allow using different scenario loaders in the future * abstracts input folder away from world * simplifies loading dataframes of different unit types a lot
