+++ assume/__init__.py
@@ -1,4 +1,5 @@
 from assume.common import MarketConfig, MarketProduct
+from assume.common.scenario_loader import load_scenario_folder
 from assume.world import World
 
 __version__ = "0.0.1"
