+++ assume/common/outputs.py
@@ -1,5 +1,4 @@
 import logging
-import os
 import shutil
 from collections import defaultdict
 from datetime import datetime
