+++ assume/common/utils.py
@@ -13,133 +13,6 @@ from assume.common.market_objects import MarketConfig, MarketProduct, Orderbook
 
 logger = logging.getLogger(__name__)
 
-freq_map = {
-    "h": rr.HOURLY,
-    "m": rr.MINUTELY,
-    "d": rr.DAILY,
-    "w": rr.WEEKLY,
-}
-
-
-def load_file(
-    path: str,
-    config: dict,
-    file_name: str,
-    index: pd.DatetimeIndex = None,
-) -> pd.DataFrame:
-    df = None
-
-    if file_name in config:
-        file_path = f"{path}/{config[file_name]}"
-    else:
-        file_path = f"{path}/{file_name}.csv"
-
-    try:
-        df = pd.read_csv(
-            file_path,
-            index_col=0,
-            encoding="utf-8",
-            na_values=["n.a.", "None", "-", "none", "nan"],
-        )
-
-        for col in df:
-            # check if the column is of dtype int
-            if df[col].dtype == "int":
-                # convert the column to float
-                df[col] = df[col].astype(float)
-
-        if index is not None:
-            if len(df.index) == 1:
-                return df
-
-            if len(df.index) != len(index):
-                logger.warning(
-                    f"Length of index does not match length of {file_name} dataframe. Attempting to resample."
-                )
-                df = attempt_resample(df, index)
-            else:
-                df.index = index
-
-        return df
-
-    except FileNotFoundError:
-        logger.warning(f"File {file_name} not found. Returning None")
-
-
-def attempt_resample(
-    df: pd.DataFrame,
-    index: pd.DatetimeIndex,
-) -> pd.DataFrame:
-    if len(df.index) > len(index):
-        temp_index = pd.date_range(start=index[0], end=index[-1], periods=len(df))
-        df.index = temp_index
-        df = df.resample(index.freq).mean()
-        logger.info("Resampling successful.")
-        return df
-    elif len(df.index) < len(index):
-        logger.warning(
-            "Index length mismatch. Upsampling not supported. Returning None"
-        )
-        return None
-
-
-def convert_to_rrule_freq(string):
-    freq = freq_map[string[-1]]
-    interval = int(string[:-1])
-    return freq, interval
-
-
-def make_market_config(
-    id: str,
-    market_params: dict,
-    world_start: datetime,
-    world_end: datetime,
-):
-    freq, interval = convert_to_rrule_freq(market_params["opening_frequency"])
-    start = market_params.get("start_date")
-    end = market_params.get("end_date")
-    if start:
-        start = pd.Timestamp(start)
-    if end:
-        end = pd.Timestamp(end)
-    start = start or world_start
-    end = end or world_end
-
-    market_products = [
-        MarketProduct(
-            duration=pd.Timedelta(product["duration"]),
-            count=product["count"],
-            first_delivery=pd.Timedelta(product["first_delivery"]),
-        )
-        for product in market_params["products"]
-    ]
-    market_config = MarketConfig(
-        name=id,
-        market_products=market_products,
-        product_type=market_params.get("product_type", "energy"),
-        opening_hours=rr.rrule(
-            freq=freq,
-            interval=interval,
-            dtstart=start,
-            until=end,
-            cache=True,
-        ),
-        opening_duration=pd.Timedelta(market_params["opening_duration"]),
-        market_mechanism=market_params["market_mechanism"],
-        maximum_bid_volume=market_params.get("maximum_bid_volume", 1e6),
-        maximum_bid_price=market_params.get("maximum_bid_price", 3000),
-        minimum_bid_price=market_params.get("minimum_bid_price", -3000),
-        maximum_gradient=market_params.get("max_gradient"),
-        volume_unit=market_params.get("volume_unit", "MW"),
-        volume_tick=market_params.get("volume_tick"),
-        price_unit=market_params.get("price_unit", "€/MWh"),
-        price_tick=market_params.get("price_tick"),
-        additional_fields=market_params.get("additional_fields", []),
-        supports_get_unmatched=market_params.get("supports_get_unmatched", False),
-    )
-
-    return market_config
-
 
 def initializer(func):
     """
