+++ examples/example_01a_sqlite.py
@@ -3,7 +3,7 @@ import logging
 import os
 from os import getenv
 
-from assume import World
+from assume import World, load_scenario_folder
 
 log = logging.getLogger(__name__)
 
@@ -18,7 +18,7 @@ if __name__ == "__main__":
     study_case = "example_01a"
 
     world = World(database_uri=DATABASE_URI, export_csv_path=EXPORT_CSV_PATH)
-    world.load_scenario(
-        inputs_path="examples/inputs", scenario=scenario, study_case=study_case
+    load_scenario_folder(
+        world, inputs_path="examples/inputs", scenario=scenario, study_case=study_case
     )
     world.run()
