+++ tests/test_utils.py
@@ -7,12 +7,11 @@ from dateutil import rrule as rr
 from dateutil.relativedelta import relativedelta as rd
 
 from assume.common.market_objects import MarketConfig, MarketProduct
+from assume.common.scenario_loader import convert_to_rrule_freq, make_market_config
 from assume.common.utils import (
     aggregate_step_amount,
-    convert_to_rrule_freq,
     get_available_products,
     initializer,
-    make_market_config,
     plot_orderbook,
 )
